requirements.txt
google-generativeai
yt-dlp
openai-whisper
ffmpeg